% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{disable_adaptive_quadrature_cpp}
\alias{disable_adaptive_quadrature_cpp}
\title{Disable adaptive quadrature}
\usage{
disable_adaptive_quadrature_cpp(fm_ptr)
}
\arguments{
\item{fm_ptr}{External pointer to FactorModel object}
}
\value{
No return value. Called for its side effect of disabling adaptive
quadrature and clearing any observation weights on the FactorModel
pointed to by \code{fm_ptr}.
}
\description{
Reverts to standard (non-adaptive) quadrature integration.
}
