% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_results.R
\name{.build_param_table}
\alias{.build_param_table}
\title{Build parameter table from result object}
\usage{
.build_param_table(result)
}
\description{
Build parameter table from result object
}
\keyword{internal}
