% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_model.R
\name{estimate_model}
\alias{estimate_model}
\title{Estimate model}
\usage{
estimate_model(ms, control)
}
\arguments{
\item{ms}{an object of class model_system}

\item{control}{an object of class estimation_control}
}
\value{
description
}
\description{
Estimate model
}
