% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evaluate_likelihood_cpp}
\alias{evaluate_likelihood_cpp}
\title{Evaluate log-likelihood for given parameters}
\usage{
evaluate_likelihood_cpp(
  fm_ptr,
  params,
  compute_gradient = FALSE,
  compute_hessian = FALSE
)
}
\arguments{
\item{fm_ptr}{External pointer to FactorModel object}

\item{params}{Vector of parameters}

\item{compute_gradient}{Whether to compute gradient (default FALSE)}

\item{compute_hessian}{Whether to compute Hessian (default FALSE)}
}
\value{
List with:
\itemize{
\item logLikelihood: scalar log-likelihood value
\item gradient: vector of length n_param_free (if requested)
\item hessian: vector of length n_param_free*(n_param_free+1)/2 stored as
upper-triangular in row-major order (if requested).
To expand to full symmetric matrix in R:
\code{idx <- 1; for(i in 1:n) for(j in i:n) { H[i,j] <- H[j,i] <- hess[idx]; idx <- idx + 1 }}
}
}
\description{
Evaluate log-likelihood for given parameters
}
