% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evaluate_loglik_only_cpp}
\alias{evaluate_loglik_only_cpp}
\title{Evaluate log-likelihood only (for optimization)}
\usage{
evaluate_loglik_only_cpp(fm_ptr, params)
}
\arguments{
\item{fm_ptr}{External pointer to FactorModel object}

\item{params}{Vector of parameters}
}
\value{
Log-likelihood value
}
\description{
Evaluate log-likelihood only (for optimization)
}
