% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{extract_free_params_cpp}
\alias{extract_free_params_cpp}
\title{Extract free parameters from full parameter vector}
\usage{
extract_free_params_cpp(fm_ptr, full_params)
}
\arguments{
\item{fm_ptr}{External pointer to FactorModel object}

\item{full_params}{Full parameter vector (size n_param)}
}
\value{
Vector of free parameters only (size n_param_free)
}
\description{
Given a full parameter vector (including fixed parameters),
extract only the free parameters based on the model's fixed parameter mask.
}
