% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gauss_hermite_quadrature}
\alias{gauss_hermite_quadrature}
\title{Compute Gauss-Hermite quadrature nodes and weights}
\usage{
gauss_hermite_quadrature(n)
}
\arguments{
\item{n}{Number of quadrature points}
}
\value{
List with nodes and weights
}
\description{
Compute Gauss-Hermite quadrature nodes and weights
}
