% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_model_component.R
\name{get_component_name}
\alias{get_component_name}
\alias{get_component_name.model_component}
\title{Get component name (internal)}
\usage{
get_component_name(x, ...)

\method{get_component_name}{model_component}(x, ...)
}
\arguments{
\item{x}{Object to extract name from (a \code{model_component}).}

\item{...}{Additional arguments (not used).}
}
\value{
A single character string with the component name, or
\code{NA_character_} if the component has no name set.
}
\description{
Get component name (internal)
}
\keyword{internal}
