% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_coefficient.R
\name{get_fixed_coefficient_value}
\alias{get_fixed_coefficient_value}
\title{Get the fixed value for a covariate's coefficient}
\usage{
get_fixed_coefficient_value(component, covariate, choice = NULL)
}
\arguments{
\item{component}{A model_component object}

\item{covariate}{Character. Name of the covariate.}

\item{choice}{Integer (optional). For multinomial logit, which choice.}
}
\value{
The fixed value, or NULL if not fixed
}
\description{
Get the fixed value for a covariate's coefficient
}
\keyword{internal}
