% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_parameter_info_cpp}
\alias{get_parameter_info_cpp}
\title{Get parameter counts from FactorModel}
\usage{
get_parameter_info_cpp(fm_ptr)
}
\arguments{
\item{fm_ptr}{External pointer to FactorModel object}
}
\value{
List with parameter count information
}
\description{
Get parameter counts from FactorModel
}
