% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_parameters.R
\name{get_second_order_loading_init}
\alias{get_second_order_loading_init}
\title{Get second-order loading initialization values and names}
\usage{
get_second_order_loading_init(comp, choice = NULL)
}
\arguments{
\item{comp}{Model component object}

\item{choice}{Integer or NULL. For multinomial logit, which choice.}
}
\value{
List with values and names for second-order loadings
}
\description{
Helper function to generate initial values and parameter names for
quadratic and interaction factor loadings.
}
\keyword{internal}
