% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initialize_factor_model_cpp}
\alias{initialize_factor_model_cpp}
\title{Initialize a FactorModel C++ object from R model system}
\usage{
initialize_factor_model_cpp(
  model_system,
  data,
  n_quad = 8L,
  init_params = NULL
)
}
\arguments{
\item{model_system}{R model_system object}

\item{data}{Data frame or matrix with all variables}

\item{n_quad}{Number of quadrature points}

\item{init_params}{Optional initial parameter vector (used to set fixed parameter values)}
}
\value{
External pointer to FactorModel object
}
\description{
Initialize a FactorModel C++ object from R model system
}
