% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_coefficient.R
\name{is_coefficient_fixed}
\alias{is_coefficient_fixed}
\title{Check if a covariate's coefficient is fixed}
\usage{
is_coefficient_fixed(component, covariate, choice = NULL)
}
\arguments{
\item{component}{A model_component object}

\item{covariate}{Character. Name of the covariate to check.}

\item{choice}{Integer (optional). For multinomial logit, which choice to check.}
}
\value{
Logical TRUE if fixed, FALSE otherwise
}
\description{
Check if a covariate's coefficient is fixed
}
\keyword{internal}
