% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_coefficient.R
\name{is_type_intercept_fixed}
\alias{is_type_intercept_fixed}
\title{Check if a type intercept is fixed}
\usage{
is_type_intercept_fixed(component, type, choice = NULL)
}
\arguments{
\item{component}{A model_component object}

\item{type}{Integer. Type number to check (2 to n_types).}

\item{choice}{Integer (optional). For multinomial logit, which choice to check.}
}
\value{
Logical TRUE if fixed, FALSE otherwise
}
\description{
Check if a type intercept is fixed
}
\keyword{internal}
