% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_coefficient.R
\name{n_fixed_coefficients}
\alias{n_fixed_coefficients}
\title{Get the number of fixed coefficients in a model component}
\usage{
n_fixed_coefficients(component)
}
\arguments{
\item{component}{A model_component object}
}
\value{
Integer count of fixed coefficients
}
\description{
Get the number of fixed coefficients in a model component
}
\keyword{internal}
