% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_coefficient.R
\name{n_fixed_type_intercepts}
\alias{n_fixed_type_intercepts}
\title{Get the number of fixed type intercepts in a model component}
\usage{
n_fixed_type_intercepts(component)
}
\arguments{
\item{component}{A model_component object}
}
\value{
Integer count of fixed type intercepts
}
\description{
Get the number of fixed type intercepts in a model component
}
\keyword{internal}
