% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_model.R
\name{prepare_model_system_for_cpp}
\alias{prepare_model_system_for_cpp}
\title{Helper function to convert model_system to format expected by C++}
\usage{
prepare_model_system_for_cpp(model_system, data)
}
\arguments{
\item{model_system}{model_system object}

\item{data}{Data frame}
}
\value{
List in format expected by initialize_factor_model_cpp
}
\description{
Helper function to convert model_system to format expected by C++
}
\keyword{internal}
